updated tree-sitter grammar to support bpf map field macros
bpf map field macros at the current projects state are unsupported. To add this a change in tree-sitter grammar was required. That is, the _field_declaration_list_item rule now has an additional choice that checks for a call_expression with a ';' as used in the bpf standard to specify map fields using the available macros. The tree can now extend its parsing depth when it sees a bpf macro call allowing us to query for __uint(a, sizeof(b)) usage.