updated tree-sitter grammar to support bpf map field macros
bpf map field macros (__uint, __type) at the current projects state are
unsupported. To add this a change in tree-sitter grammar was required. That is,
the _field_declaration_list_item rule now has an additional choice that checks
for a `type_field_declarator_macro` or a `uint_field_declarator_macro` which
are defined in grammar.js. Unit tests for these two rules have been added to
/test/corpus/bpf.txt.